t/data/parsing/temma-node-1.dat

#data
<t:element/><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma element>
|     <link>

#data
axs<t:element/><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "axs"
|     <temma element>
|     <link>

#data
<t:element>a</t:elemenT><t:element/><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma element>
|       "a"
|     <temma element>
|     <link>

#data
<t:element><p>foo</t:element><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma element>
|       <p>
|         "foo"
|     <link>

#data
<t:element><p>f</t:element><p>b
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma element>
|       <p>
|         "f"
|   <body>
|     <p>
|       "b"

#data
<t:attr/><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <temma attr>
|   <head>
|     <link>

#data
axs<t:attr/><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "axs"
|     <temma attr>
|     <link>

#data
<t:attr>a</t:attr><t:attr/><link/>
#errors
1;19;m;unmatched end tag;;t:attr
#document
| <!DOCTYPE html>
| <html>
|   <temma attr>
|   <body>
|     "a"
|     <temma attr>
|     <link>

#data
<t:attr><p>foo</t:attr><link/>
#errors
1;24;m;unmatched end tag;;t:attr
#document
| <!DOCTYPE html>
| <html>
|   <temma attr>
|   <body>
|     <p>
|       "foo"
|       <link>

#data
<t:attr><p>f</t:attr><p>b
#errors
1;22;m;unmatched end tag;;t:attr
#document
| <!DOCTYPE html>
| <html>
|   <temma attr>
|   <body>
|     <p>
|       "f"
|     <p>
|       "b"

#data
<meta><t:attr/><base/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <meta>
|     <temma attr>
|     <base>

#data
<meta><t:attr><link>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <meta>
|     <temma attr>
|     <link>

#data
<html><t:attr name='"manifest"' value='1243'/><head>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <temma attr>
|     name=""manifest""
|     value="1243"
|   <head>

#data
<t:comment/><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma comment>
|     <link>

#data
axs<t:comment/><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "axs"
|     <temma comment>
|     <link>

#data
<t:comment>a</t:comment><t:comment/><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma comment>
|       "a"
|     <temma comment>
|     <link>

#data
<t:comment><p>foo</t:comment><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma comment>
|       <p>
|         "foo"
|     <link>

#data
<t:comment><p>f</t:comment><p>b
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma comment>
|       <p>
|         "f"
|   <body>
|     <p>
|       "b"

#data
<svg><t:element><g/></t:element></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <temma element>
|         <svg g>

#data
<t:text>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma text>

#data
a<t:text>bbb
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "a"
|     <temma text>
|     "bbb"

#data
a<t:text>b</t:text>bb
#errors
1;20;m;unmatched end tag;;t:text
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "a"
|     <temma text>
|     "bbb"

#data
<t:attr><t:text>bbb
#errors
#document
| <!DOCTYPE html>
| <html>
|   <temma attr>
|   <body>
|     <temma text>
|     "bbb"

#data
<t:text t:parse>hoge</t:text>xxy
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma text>
|       temma parse=""
|       "hoge"
|     "xxy"

#data
<title t:parse><t:text></title><t:text>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <title>
|       temma parse=""
|       <temma text>
|   <body>
|     <temma text>

#data
<t:class><!---->
#errors
#document
| <!DOCTYPE html>
| <html>
|   <temma class>
|   <!--  -->

#data
<link><t:class><base>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <link>
|     <temma class>
|     <base>

#data
<p><t:class>hoge
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <p>
|       <temma class>
|       "hoge"

#data
<title><t:class>foo</title>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <title>
|       "<t:class>foo"

#data
<svg><t:attr><t:attr><g></g></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <temma attr>
|       <temma attr>
|       <svg g>

#data
<math><t:attr><t:attr><g></g></math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <temma attr>
|       <temma attr>
|       <math g>

#data
<svg><foreignobject><t:attr></foreignobject></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         <temma attr>

#data
<t:barehtml>hoge
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <temma barehtml>
|     "hoge"

#data
<title t:parse><t:barehtml>hoge</title>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <title>
|       temma parse=""
|       <temma barehtml>
|       "hoge"
